refactor: request 객체 패키지 분리
