+++ src/main/java/kitchenpos/menus/tobe/application/dto/MenuGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.application;
+package kitchenpos.menus.tobe.application.dto;
 
 import java.util.Objects;
