+++ src/main/java/kitchenpos/menus/tobe/application/dto/MenuRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.application;
+package kitchenpos.menus.tobe.application.dto;
 
 import kitchenpos.menus.tobe.domain.MenuProduct;
