+++ src/main/java/kitchenpos/menus/tobe/ui/MenuRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.tobe.ui;
 
-import kitchenpos.menus.tobe.application.MenuPriceRequest;
-import kitchenpos.menus.tobe.application.MenuRequest;
+import kitchenpos.menus.tobe.application.dto.MenuPriceRequest;
+import kitchenpos.menus.tobe.application.dto.MenuRequest;
 import kitchenpos.menus.tobe.application.MenuService;
 import kitchenpos.menus.tobe.domain.Menu;
 import org.springframework.http.ResponseEntity;
